requirements.txt
Flask
PyPDF2
python-docx
nltk
scikit-learn
numpy
pandas
matplotlib
pymupdf